Make killopenssh.sh more robust

+++ ext/tls/axTLS/ssl/test/killopenssl.sh
@@ -1,3 +1,5 @@
 #!/bin/sh
-awk '{print $1}' "../ssl/openssl.pid" | xargs kill -9
-rm -f ../ssl/openssl.pid
+if [ -f "../ssl/openssl.pid" ]; then
+    awk '{print $1}' "../ssl/openssl.pid" | xargs kill -9
+    rm -f ../ssl/openssl.pid
+fi
